Extract util function into utils file
